Rename "inject" package to "dagger"
