+++ sso/src/test/java/drewhamilton/skylight/sso/dagger/SsoSkylightComponentTest.kt
@@ -1,4 +1,4 @@
-package drewhamilton.skylight.sso.inject
+package drewhamilton.skylight.sso.dagger
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
